+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.1.1
   - Fix remove-service #1
+  - Add python and PyYAML
+  - Fix locales
+  - Fix my_init
 
 ## 0.1.0
   - Initial release